cleaning up dev branch for merge
